requirements.txt
streamlit>=1.21,<1.23
SQLAlchemy>=2.0
pymysql>=1.1
requests>=2.28
python-dotenv>=1.0
pandas>=2.0
